+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 3.0.0
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Switch to latest techdivision/import-attribute version 7.0.*
+* Added functionality to copy parent attribute set attribute groups + attribute relations
+
 # Version 2.0.0
 
 ## Bugfixes